finish agora page refactor
